+++ Makefile
@@ -10,7 +10,7 @@ INCLUDE_DIR = ./include
 MAIN = main signal strarrutils
 CMD = cd echo env exec exec_proc exit export pwd unset
 PARSER =  arg_splitter asterisk_splitter expand_envvar expand_wildcard get_arg_num get_file_list get_length is load_lst parser rm_quotes
-REDIR = cmd_redir exec_redir utils_redir
+REDIR = cmd_redir exec_redir utils_redir utils_redir2
 PIPE = utils_pipe utils_pipe2 utils_pipe3
 
 SRCS =	$(addsuffix .c, $(MAIN)) \
