+++ include/minishell.h
@@ -93,6 +93,8 @@ void	argu_left_change(char **args);
 int		redir_excute(char **args, t_node *node);
 int		repeat_check(char **args, t_node *node);
 void	original_store(char **args, t_node *node);
+void	args_left_move(char **args, int i);
+int		print_err(char **args, int i);
 // pipe
 void	exec_child(char **args, char **envp, t_node *node);
 void	exec_parents(int pid, char **args, char **envp, t_node *node);
